Move components shared between landing page and app to shared dir
